requirements.txt
Flask
pymysql
requests
python-dotenv
Jinja2
MarkupSafe